+++ app/(private)/dashboard/meetup/[meetupId]/contacts/new/page.tsx
@@ -3,7 +3,7 @@ import Link from "next/link";
 import { getTags } from "../_server/server";
 
 import { ContactsErrorCard } from "@/components/contacts/contacts-error-card";
-import { ContactForm } from "@/components/contacts/form/ContactForm";
+import { CreateContactForm } from "@/components/contacts/form/CreateContactForm";
 import { DashboardHeader } from "@/components/dashboard/dashboard-header";
 import { Button } from "@/components/ui/button";
 import { routes } from "@/util/routes";
@@ -50,7 +50,7 @@ export default async function CreateContacts({
           </Button>
         </div>
       </div>
-      <ContactForm meetupId={meetupId} tags={tags.data} />
+      <CreateContactForm meetupId={meetupId} tags={tags.data} />
     </div>
   );
 }
